Split the data from the implementation.
